env2/requirements.txt
Babel==2.10.1
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
httplib2==0.20.4
idna==3.3
numpy==1.21.6
oauthlib==3.2.0
pandas==1.3.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tkcalendar==1.6.1
uritemplate==4.1.1
urllib3==1.26.9
